backend/requirements.txt
pygithub
langchain
langchain-community
langchain-pinecone
langchain-huggingface
pinecone-client
tiktoken
openai
python-dotenv
sentence-transformers
tree-sitter
tree-sitter-python
tree-sitter-java
tree-sitter-cpp
tree-sitter-javascript
tree-sitter-go
tree-sitter-html